Optimized Objects.sizeof() method using new ByteCountingOutputStream.
git-svn-id: https://svn.apache.org/repos/asf/incubator/wicket/trunk@458752 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket/src/java/wicket/Page.java
@@ -491,7 +491,7 @@ public abstract class Page extends MarkupContainer implements IRedirectListener,
 	/**
 	 * @return Size of this page in bytes
 	 */
-	public final int getSizeInBytes()
+	public final long getSizeInBytes()
 	{
 		this.pageMap = null;
 		return Objects.sizeof(this);
